Overloaded addTimedSpawn()
The start of 3.9.7.10-SNAPSHOT

public -> private addTimedSpawn(Long index, TimedSpawn s);
public addTimedSpawn(TimedSpawn s);

The new method no longer takes an index parameter because the index can be
calculated. The only reason to pass an index is to replace that index with
a new TimedSpawn.

The old method was temporarily changed to private in order to generate
errors (on purpose) in order to ascertain WHERE the method is being
used... And if possible, update any uses to the new method call.

Also, the "addspawn" command will be easier for Admins to use because they
won't have to pass an index.

Will need a new command "setspawn" which takes an index... So that any
spawns can be replaced with new ones.

To-Do:

Add a new API method for Arena that will replace a TimedSpawn.

Currently, the only way to achieve this is via

Arena.getTimedSpawns() which returns a Map<Long, TimedSpawn>.

And it's not immediately clear to developers if this map can be modified
in order to update any TimedSpawn or if the map is just a cloned copy.

In order to finish 3.9.7.10
1. All the compile time IDE errors should be fixed.
2. The "setspawn" command should be implemented.
3. Also,

ideally, since timedSpawn is a field of Arena.java, then it should be
passed into the constructor (Dependency Injection will solve all our
problems).

As it stands right now, the ArenaEditorExecutor() will perform an
arena.getTimedSpawns() and then do a null check... One should not have to
perform a null check after a getter() method... Especially when we can
return an empty, new HashMap<>()... The end-user can then safely dispatch
methods of that new object without fear of generating a
NullPointerException.