Adds requested feature #304 /sb command.
This feature was requested by Adambor in order to use with the
VariableTriggers plugin and his custom scripts (to create custom arenas).

There probably isn't a use-case-scenario for the `/sb undo` cmd.
I just wanted to make one, but...

The undo command probably needs to be changed to specify an undo operation
for a particular match... Meaning that it will require an Arena parameter
and that a stack of commands will exist for each arena.

It also means that when this particular match ends, the stack can be
cleared.

Right now, the stack size is limited to 10.

`/sb unregister` cmd is only partially implemented.
The match will no longer have an `amount2win`, but the creation label will
persist.