Made output optional for ScoreboardExecutor.java
A user complained about the output spamming the console and server logs,
so I made the output optional.

New commands for quiet mode, verbose mode, and to toggle between the modes

./sb quiet
./sb verbose
./sb toggle

Verbose mode is the default. But I also added a new configuration option
in the config.yml to change this.