BattleArena v3.9.9.9
Fixes BattleArena.java line 135: "McMMOConfig"

The file was named mcMMOConfig.

So there was two possible solutions:
Either
1. change the string delcaration on line 135 to "mcMMOConfig"
2. Or change the file name to "McMMOConfig" to match the string.

And actually, moving /default_files/config.yml to a new location has
caused the exact same NPE on a different line.

So we need to either move the config.yml back to the original location,
or point the code to the new location.

And also, with the change in location, config.yml is not even being
included in the build. So the pom.xml needs to be updated.