Prevent player exploit with negative numbers.
./<arenaType> duel <player> rake=-100

A negative rake could potentially be abused and exploited by players. The
above command would essentially DOUBLE the amount of money won.

Only negative values for the command are prevented.

I believe it is possible for Admins to configure a negative rake (as a
kind of bonus) in /plugins/BattleArena/config.yml

node:
"defaultOptions.defaultDuelOptions:"

list value: [rake=0.0]