Added DuelOption.RAKE
Feature Requested by werwew19.

A RAKE (in poker) is a portion of the pot that is kept by the house.

In other words, a RAKE can be thought of as a tax, fee, or money sink.

So, it's a small fee for betting money.

The RAKE is a percentage.
This means that 5.0 = 5%

So for a bet of $100 each & two players, the winner would receive 95%
($190).