Changed MatchMessager
field MatchMessager mc is no longer final.

This is to make it easier for those using the BattleArena API to inject
their own MatchMessager.

Along the same lines, I think it would be good to add more information
to the MatchMessageEvent (to encourage users to use an Event, rather than
injecting their own implementation). Because, as it stands right now, no
actual MatchMessageEvent is being fired, except for
MatchIntervalMessageEvent and MatchTimeExpiredMessageEvent... So listening
to this event provides limited functionality.

I think it would be good to include the message Node with the
MatchMessageEvent, so that listeners can check for a specific node being
fired, and customize accordingly... This would mean that BattleArena would
have to fire the MatchMessageEvent for every message... allowing listeners
to change the message.

All this comes from a question in IRC for how to customize Match messages
at runtime (rather than customizing via configuration file). I think
calling MatchMessageEvent more would be the appropriate solution to this
problem (altho, right now, the only solution is to inject your own
implementation of MatchMessager into the Match object).