Fixes #286 chest loading.
https://github.com/alkarinv/BattleArena/issues/286

"giveItems" was changed to "items"

The reason why this was done was because if I did the opposite and changed
"items" to "giveItems", then yes, I could check arenas.yml chest spawns
for the existence of either node... however, consider someone
automatically updates BattleArena, uses the command to create a chest
spawn (which in this hypothetical example, would save under the yaml node
"giveItems:"). Now consider that they that (for whatever reason) revert
back to an older version of BattleArena... that's a potential problem that
I wanted to avoid. The problem would be that older versions of BA
would be looking for an "items:" node, not a "giveItems:" node.