Register & handle onKill
onKill was registered in ListenerAdder.addlisteners()

The idea behind ListenerAdder is to only register an event listener if an
arena is configured with that particular option, in this case, the onKill
option.

Modified ArenaPlayerKillListener.java

Originally, the primary purpose of this case was to listen for arena kills
and forward them to the BattleTracker database... But this appears to be
an un-implemented feature, as this class was never registered, nor was the
method annotated.

Right now, only the new onKill() method is annotated to receive events.
onArenaPlayerKillEvent() is still not annotated, nor is it receiving
events. This is intended... until BattleTracker can be updated.

The two responsibilities are separated so that if an exception occurs, it
will not effect the other:

1. Forward kills to BattleTracker
2. Give rewards onKill

These two responsibilities should be separated into their own class in the
future.