v3.9.10.15 addresses issue #69
with custom hats

There appears to be two possible reasons why custom hats aren't being
restored:

1. Possible NPE on line 431 that was never reported.
2. Arena lacks - clearInventory option and is mis-configured.

SELF-NOTE:
The method spec according to the comments do not seem to match up with the
code. Also, this method addArmorToInventory() appears to be wrong in so
many ways: specifically & most importantly on line 459 where a non-empty
slot occurs. When old is added, it's not being removed & new isn't being
added. When new is added, old isn't being added (according to the spec).